third_party/blink/web_tests/virtual/disable-ua-client-hint/external/wpt/client-hints/sandbox/iframe-csp.https-expected.txt

This is a testharness.js-based test.
[FAIL] CSP sandboxed iframe does not send client hint headers
  assert_equals: message from opened frame expected "PASS" but got "FAIL sec-ch-ua True None"
Harness: the test ran to completion.
